+++ chatService/src/main/java/de/fhac/lsicc/chatservice/dto/ChatDTO.java
@@ -5,5 +5,5 @@ import java.util.List;
 /**
  * Created by Lukas Szimtenings on 7/18/2022.
  */
-public record ChatDTO(String chatName, Integer chatId, List<Integer> participants) {
+public record ChatDTO(String chatName, Long chatId, List<Long> participants) {
 }
