+++ chatService/src/main/java/de/fhac/lsicc/chatservice/repsitories/ChatRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.repository.CrudRepository;
 /**
  * Created by Lukas Szimtenings on 7/18/2022.
  */
-public interface ChatRepository extends CrudRepository<Chat, Integer>
+public interface ChatRepository extends CrudRepository<Chat, Long>
 {
 }
